Update MSToolkit to address raw file parsing issues.
git-svn-id: https://svn.code.sf.net/p/comet-ms/code@984 299d2d02-8f3b-485e-a07e-71e8e5e16577

+++ trunk/comet-ms/MSToolkit/src/mzParser/PWIZface.cpp
@@ -61,7 +61,7 @@ bool ChromatogramList::get() {
 	return true;
 }
 
-unsigned int ChromatogramList::size() {
+size_t ChromatogramList::size() {
 	if(vChromatIndex==NULL) {
 		cerr << "Get chromatogram list first." << endl;
 		return 0;
